+++ wger/settings_global.py
@@ -99,7 +99,9 @@ BOWER_INSTALLED_APPS = (
     'DataTables',
     'components-font-awesome',
     'tinymce',
-    'metrics-graphics'
+    'metrics-graphics',
+    'devbridge-autocomplete#1.2.x',
+#    'sortablejs#1.4.x',
 )
 
 
@@ -133,26 +135,40 @@ AUTHENTICATION_BACKENDS = (
     'wger.utils.helpers.EmailAuthBackend'
 )
 
-# Set the context processors
-TEMPLATE_CONTEXT_PROCESSORS = (
-    'wger.utils.context_processor.processor',
-    'django.contrib.auth.context_processors.auth',
-    'django.contrib.messages.context_processors.messages',
+TEMPLATES = [
+    {
+        'BACKEND': 'django.template.backends.django.DjangoTemplates',
+        # 'APP_DIRS': True,
+        'OPTIONS': {
+            'context_processors': [
+                'wger.utils.context_processor.processor',
 
-    # Django mobile
-    'django_mobile.context_processors.flavour',
+                # Django
+                'django.contrib.auth.context_processors.auth',
+                'django.template.context_processors.debug',
+                'django.template.context_processors.i18n',
+                'django.template.context_processors.media',
+                'django.template.context_processors.static',
+                'django.template.context_processors.tz',
+                'django.contrib.messages.context_processors.messages',
 
-    # Breadcrumbs
-    'django.core.context_processors.request'
-)
+                # Django mobile
+                'django_mobile.context_processors.flavour',
 
-TEMPLATE_LOADERS = (
-    # Django mobile
-    'django_mobile.loader.Loader',
+                # Breadcrumbs
+                'django.core.context_processors.request'
+            ],
+            'loaders': [
+                # Django mobile
+                'django_mobile.loader.Loader',
 
-    'django.template.loaders.filesystem.Loader',
-    'django.template.loaders.app_directories.Loader',
-)
+                'django.template.loaders.filesystem.Loader',
+                'django.template.loaders.app_directories.Loader',
+            ],
+            'debug': False
+        },
+    },
+]
 
 # Store the user messages in the session
 MESSAGE_STORAGE = 'django.contrib.messages.storage.session.SessionStorage'
@@ -350,5 +366,6 @@ WGER_SETTINGS = {
     'USE_RECAPTCHA': False,
     'REMOVE_WHITESPACE': False,
     'ALLOW_REGISTRATION': True,
-    'EMAIL_FROM': 'wger Workout Manager <wger@example.com>'
+    'EMAIL_FROM': 'wger Workout Manager <wger@example.com>',
+    'TWITTER': False
 }